MAAS: Enable configure on MAAS deployments
Configure is now specialized for local and maas deployments.

Maas deployments will fetch the external nic from maas. It will look for
the first nic tagged with `compute`.
