+++ sunbeam-python/sunbeam/provider/local/commands.py
@@ -14,6 +14,8 @@
 # limitations under the License.
 
 import logging
+import os
+import shutil
 from pathlib import Path
 from typing import Tuple, Type
 
@@ -37,7 +39,14 @@ from sunbeam.commands.clusterd import (
     ClusterUpdateJujuControllerStep,
     ClusterUpdateNodeStep,
 )
-from sunbeam.commands.configure import SetLocalHypervisorOptions
+from sunbeam.commands.configure import (
+    DemoSetup,
+    SetHypervisorCharmConfigStep,
+    TerraformDemoInitStep,
+    UserOpenRCStep,
+    UserQuestions,
+    retrieve_admin_credentials,
+)
 from sunbeam.commands.hypervisor import (
     AddHypervisorUnitsStep,
     DeployHypervisorApplicationStep,
@@ -79,13 +88,14 @@ from sunbeam.commands.sunbeam_machine import (
     DeploySunbeamMachineApplicationStep,
     RemoveSunbeamMachineStep,
 )
-from sunbeam.commands.terraform import TerraformInitStep
+from sunbeam.commands.terraform import TerraformHelper, TerraformInitStep
 from sunbeam.jobs.checks import (
     DaemonGroupCheck,
     JujuSnapCheck,
     LocalShareCheck,
     SshKeysConnectedCheck,
     SystemRequirementsCheck,
+    VerifyBootstrappedCheck,
     VerifyFQDNCheck,
     VerifyHypervisorHostnameCheck,
 )
@@ -105,11 +115,13 @@ from sunbeam.jobs.common import (
     validate_roles,
 )
 from sunbeam.jobs.deployment import Deployment
-from sunbeam.jobs.juju import CONTROLLER, JujuHelper
+from sunbeam.jobs.juju import CONTROLLER, JujuHelper, ModelNotFoundException, run_sync
 from sunbeam.jobs.manifest import AddManifestStep, Manifest
 from sunbeam.provider.base import ProviderBase
 from sunbeam.provider.local.deployment import LOCAL_TYPE, LocalDeployment
+from sunbeam.provider.local.steps import LocalSetHypervisorUnitsOptionsStep
 from sunbeam.utils import CatchGroup
+from sunbeam.versions import TERRAFORM_DIR_NAMES
 
 LOG = logging.getLogger(__name__)
 console = Console()
@@ -131,12 +143,18 @@ class LocalProvider(ProviderBase):
         """A local provider cannot add deployments."""
         pass
 
-    def register_cli(self, init: click.Group, deployment: click.Group):
+    def register_cli(
+        self,
+        init: click.Group,
+        configure: click.Group,
+        deployment: click.Group,
+    ):
         """Register local provider commands to CLI.
 
         Local provider does not add commands to the deployment group.
         """
         init.add_command(cluster)
+        configure.add_command(configure_cmd)
         cluster.add_command(bootstrap)
         cluster.add_command(add)
         cluster.add_command(join)
@@ -572,7 +590,7 @@ def join(
                 AddHypervisorUnitsStep(
                     client, name, jhelper, deployment.infrastructure_model
                 ),
-                SetLocalHypervisorOptions(
+                LocalSetHypervisorUnitsOptionsStep(
                     client,
                     name,
                     jhelper,
@@ -677,3 +695,126 @@ def remove(ctx: click.Context, name: str, force: bool) -> None:
         f"Run command 'sudo /sbin/remove-juju-services' on node {name} "
         "to reuse the machine."
     )
+
+
+@click.command("deployment")
+@click.pass_context
+@click.option("-a", "--accept-defaults", help="Accept all defaults.", is_flag=True)
+@click.option(
+    "-m",
+    "--manifest",
+    help="Manifest file.",
+    type=click.Path(exists=True, dir_okay=False, path_type=Path),
+)
+@click.option(
+    "-o",
+    "--openrc",
+    help="Output file for cloud access details.",
+    type=click.Path(dir_okay=False, path_type=Path),
+)
+def configure_cmd(
+    ctx: click.Context,
+    openrc: Path | None = None,
+    manifest: Path | None = None,
+    accept_defaults: bool = False,
+) -> None:
+    deployment: Deployment = ctx.obj
+    client = deployment.get_client()
+    preflight_checks = []
+    preflight_checks.append(DaemonGroupCheck())
+    preflight_checks.append(VerifyBootstrappedCheck(client))
+    run_preflight_checks(preflight_checks, console)
+
+    # Validate manifest file
+    manifest_obj = None
+    if manifest:
+        manifest_obj = Manifest.load(
+            deployment, manifest_file=manifest, include_defaults=True
+        )
+    else:
+        manifest_obj = Manifest.load_latest_from_clusterdb(
+            deployment, include_defaults=True
+        )
+
+    LOG.debug(
+        f"Manifest used for deployment - preseed: {manifest_obj.deployment_config}"
+    )
+    LOG.debug(
+        f"Manifest used for deployment - software: {manifest_obj.software_config}"
+    )
+    preseed = manifest_obj.deployment_config or {}
+
+    name = utils.get_fqdn()
+    tfplan = "demo-setup"
+    tfplan_dir: str = TERRAFORM_DIR_NAMES.get(tfplan)
+    snap = Snap()
+    manifest_tfplans = manifest_obj.software_config.terraform
+    src = manifest_tfplans.get(tfplan).source
+    dst = snap.paths.user_common / "etc" / deployment.name / tfplan_dir
+    try:
+        os.mkdir(dst)
+    except FileExistsError:
+        pass
+    # NOTE: install to user writable location
+    LOG.debug(f"Updating {dst} from {src}...")
+    shutil.copytree(src, dst, dirs_exist_ok=True)
+
+    jhelper = JujuHelper(deployment.get_connected_controller())
+    try:
+        run_sync(jhelper.get_model(OPENSTACK_MODEL))
+    except ModelNotFoundException:
+        LOG.error(f"Expected model {OPENSTACK_MODEL} missing")
+        raise click.ClickException("Please run `sunbeam cluster bootstrap` first")
+    admin_credentials = retrieve_admin_credentials(jhelper, OPENSTACK_MODEL)
+    tfhelper = TerraformHelper(
+        path=snap.paths.user_common / "etc" / deployment.name / tfplan_dir,
+        env=admin_credentials,
+        plan=tfplan,
+        backend="http",
+        clusterd_address=deployment.get_clusterd_http_address(),
+    )
+    answer_file = tfhelper.path / "config.auto.tfvars.json"
+    plan = [
+        JujuLoginStep(deployment.juju_account),
+        UserQuestions(
+            client,
+            answer_file=answer_file,
+            deployment_preseed=preseed,
+            accept_defaults=accept_defaults,
+        ),
+        TerraformDemoInitStep(client, tfhelper),
+        DemoSetup(
+            client=client,
+            tfhelper=tfhelper,
+            answer_file=answer_file,
+        ),
+        UserOpenRCStep(
+            client=client,
+            tfhelper=tfhelper,
+            auth_url=admin_credentials["OS_AUTH_URL"],
+            auth_version=admin_credentials["OS_AUTH_VERSION"],
+            openrc=openrc,
+        ),
+        SetHypervisorCharmConfigStep(
+            client,
+            jhelper,
+            ext_network=answer_file,
+            model=deployment.infrastructure_model,
+        ),
+    ]
+    node = client.cluster.get_node_info(name)
+
+    if "compute" in node["role"]:
+        plan.append(
+            LocalSetHypervisorUnitsOptionsStep(
+                client,
+                name,
+                jhelper,
+                deployment.infrastructure_model,
+                # Accept preseed file but do not allow 'accept_defaults' as nic
+                # selection may vary from machine to machine and is potentially
+                # destructive if it takes over an unintended nic.
+                deployment_preseed=preseed,
+            )
+        )
+    run_plan(plan, console)
